+++ src/wxMaximaArtProvider.cpp
@@ -66,6 +66,8 @@
 #include "art/config/options.h"
 #include "art/config/styles.h"
 #include "art/config/view-refresh.h"
+#include "art/config/accessibility.h"
+#include "art/config/ai-chat.h"
 
 // Used to gunzip the (gzip compressed) SVG in Memory
 wxString wxMaximaArtProvider::gunzip(unsigned char * svg_gz, size_t svg_gz_size)
@@ -138,6 +140,10 @@ wxBitmapBundle wxMaximaArtProvider::CreateBitmapBundle(const wxArtID& id,
     return wxBitmapBundle::FromSVG(wxMaximaArtProvider::gunzip(STYLES_SVG_GZ, STYLES_SVG_GZ_SIZE).mb_str(), artsize);
   } else if (id == wxmaximaART_CONFIG_VIEW_REFRESH) {
     return wxBitmapBundle::FromSVG(wxMaximaArtProvider::gunzip(VIEW_REFRESH_SVG_GZ, VIEW_REFRESH_SVG_GZ_SIZE).mb_str(), artsize);
+  } else if (id == wxmaximaART_CONFIG_ACCESSIBILITY) {
+    return wxBitmapBundle::FromSVG(wxMaximaArtProvider::gunzip(ACCESSIBILITY_SVG_GZ, ACCESSIBILITY_SVG_GZ_SIZE).mb_str(), artsize);
+  } else if (id == wxmaximaART_CONFIG_AI_CHAT) {
+    return wxBitmapBundle::FromSVG(wxMaximaArtProvider::gunzip(AI_CHAT_SVG_GZ, AI_CHAT_SVG_GZ_SIZE).mb_str(), artsize);
   } else {
     return wxNullBitmap;
   }